requirements.txt
Flask
pandas
biopython
numpy
